Added code to handle summer quarter.
During the regular school year, Dancebreak sessions occur in
Roble Gym, and studio reservations must be made on a weekly basis
(and are not confirmed until Friday of the week prior).

During summer, this doesn't apply:  sessions occur outside Roble
Gym, and reservations are confirmed prior to being added to the
locations_* table.

*  Added variable to indicate whether it is summer quarter.
   conditional on whether it is summer quarter.
*  Made the footnote about future dates being unconfirmed
   conditional on whether it is summer quarter.
*  Turned off the indication that a date is unconfirmed if it is
   summer quarter.
*  Added comments about summer quarter with new code.