Fixed white space to conform to new style guide.
(This is preparatory to the checking in of a style-guide git
pre-commit script, which will enforce this style and prevent
changes violating it from being committed.)

To help webmasters who have little-to-no experience working with
Linux and text editors, all tab characters have been converted to
space characters.  This eliminates the need for the additional
"how-to" step of setting up the webmaster's text editor to
interpret tabs as the appropriate number of space characters.

The idea is to make it easier for inexperienced webmasters to do
their jobs by requiring they perform fewer tasks, particularly
"setup" ones (if there are many, then people may be dissuaded
from volunteering to become webmasters in the first place).

Also:

*  Removed trailing space characters (per style-guide).

*  Unified block indentation to 2 spaces per level.