Updated HOW-TOs for specific GitLab requirement.
During the creation/updating of the "staging" and "main" copies
of the website, only the GitLab username and password are needed.
More specifically, no SSH keys are required for this step.  This
is now explicitly specified in the relevant HOW-TO files, as is
the warning not to upload personal SSH keys to rice.